Fixed bug where TLE data would sometimes not be found
If retrieving TLE data failed, this caused an empty file to be used for getting TLE information. Because it failed, the Python script would crash. The new Bash script prevents old files from being removed if the retreiving failed. This ideally needs to be done in the Python script, but the new Bash script adds robustness.